requirements.txt
facebook-sdk==3.1.0
python-dotenv==1.0.0
schedule==1.2.0
# pillow==9.5.0
python-dateutil==2.8.2 